blog/requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
django-debug-toolbar==3.7.0
django-debug-toolbar-request-history==0.1.4
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docker==6.0.1
drf-spectacular==0.24.2
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
packaging==21.3
psycopg2==2.9.5
psycopg2-binary==2.9.5
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.2
pytz==2022.6
PyYAML==6.0
requests==2.28.1
sqlparse==0.4.3
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.13
websocket-client==1.4.2
